Statically initialize some registers to remove warnings
Set a default value for a bunch registers. This significantly reduces the amount
of warnings spewed by ghdl from the IEEE libraries all over the place.

Interestingly this *does* affect Vivado synthesis... a small reduction in
slice LUTs and increase in registers. Timing is jiggled a bit but not in a
significant way as far as I can tell.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>