caches: Use the BRAM helper in sim
This makes the caches use the BRAM helper in sim instead of open-coded
memories. This avoids a few problems with GHDL, notably:

 - Creating wave files without option files to reduce the traced signals
is horrendous as the complete RAMs are dumped

 - GHDL uses a byte per bit for typical RAMs, using a lot more memory
to the point that it can occasionally fail allocations unless
--max-stack-alloc is used.

Additionally, it looks like this approach is a tiny bit faster.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>