intercon: Add layer of buffering between caches and arbiter
This adds a wb_buffer object which adds a cycle latency on wishbone
in both directions (todo: investigate making it downstream only) and
has a stash buffer to be able to break long stall chains.

We insert this buffer between the caches (and wb debug) and the arbiter
in the hope that it will allow the synthesizer more wiggle room in
placing the various core elements and still make timing.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>